+++ quarkus/datasource/src/main/java/io/openk9/datasource/graphql/DatasourceGraphqlResource.java
@@ -70,7 +70,7 @@ public class DatasourceGraphqlResource {
 	}
 
 	@Mutation
-	public Uni<Datasource> deleteDatasource(int datasourceId) {
+	public Uni<Datasource> deleteDatasource(long datasourceId) {
 		return datasourceService.deleteById(datasourceId);
 	}
