+++ quarkus/datasource/src/main/java/io/openk9/datasource/graphql/EntityIndexGraphqlResource.java
@@ -70,7 +70,7 @@ public class EntityIndexGraphqlResource {
 	}
 
 	@Mutation
-	public Uni<EntityIndex> deleteEntityIndex(int entityIndexId) {
+	public Uni<EntityIndex> deleteEntityIndex(long entityIndexId) {
 		return entityIndexService.deleteById(entityIndexId);
 	}
